==> Notebooks/03_feature_engineering.ipynb <==
import pandas as pd
from scipy.sparse import csr_matrix

df = pd.read_csv("../data/cleaned_book_ratings.csv")
print("Shape:", df.shape)
df.head()

# Feature Engineering

## User-level Features

# --- Favorite Author ---
# Compute average rating per user-author pair
user_author_mean = df.groupby(["user_id", "author"])["book_rating"].mean().reset_index()

# For each user, find the author with the max average rating
fav_author = user_author_mean.loc[user_author_mean.groupby("user_id")["book_rating"].idxmax()]

# Keep only user_id and author
fav_author = fav_author[["user_id", "author"]].rename(columns={"author": "fav_author"}).reset_index(drop=True)
fav_author

# --- Favorite Publisher ---

# Count number of books rated by each user-publisher pair
user_publisher_count = df.groupby(["user_id", "publisher"])["book_rating"].count().reset_index(name="count")

# For each user, get the publisher with the max count
fav_publisher = user_publisher_count.loc[user_publisher_count.groupby("user_id")["count"].idxmax()]

# Keep only user_id and publisher
fav_publisher = fav_publisher[["user_id", "publisher"]].rename(columns={"publisher": "fav_publisher"}).reset_index(drop=True)
fav_publisher

## Books-level-Features

# --- average rating ---
avg_book_rating=df.groupby('isbn')['book_rating'].mean().reset_index(name='avg_book_rate')
avg_book_rating

num_rating=df[['isbn','num_of_rating']].drop_duplicates()
book_info=avg_book_rating.merge(num_rating,on='isbn',how='inner')
book_info


### Popularity / Weighted Score (IMDb-style)

**Formula:**


$$
\text{score} = \frac{v}{v+m}R + \frac{m}{v+m}C
$$

Where:
- **R** = average rating for the book  
- **v** = number of ratings for the book  
- **m** = minimum ratings threshold   
- **C** = global average rating across all books  

**How it works:**
- If a book has **few ratings**, the score leans closer to **C** (the overall average).  
- If a book has **many ratings**, the score leans closer to its true **R**.  

👉 This prevents books with just a few perfect ratings from unfairly dominating the ranking.


print(book_info['num_of_rating'].describe())
m = 13     # minimum ratings threshold
# - Chosen m as 13 because from our dataset stats, 75% of books have <= 13 ratings
# - This ensures that books with very few ratings do not dominate the weighted score

C=df['book_rating'].mean()  # global average rating across all books  

book_info['weighted_score']=book_info['num_of_rating']*book_info['avg_book_rate']/(book_info['num_of_rating']+m) + m*C/(book_info['num_of_rating']+m)
book_info

# New Dataset

df_plus=df

df_plus=df_plus.merge(fav_author,on='user_id',how='left').merge(fav_publisher,on='user_id',how='left')
df_plus

del book_info['num_of_rating'] # df_plus contain num_of_raging column so no need to merge it

df_plus=df_plus.merge(book_info,on='isbn',how='left')
df_plus

df_plus.to_csv('../data/cleaned_book_ratings_plus.csv')

### 🛠 Feature Engineering: New Columns Added

After feature engineering, the following new features have been added to the dataset:

1. **fav_author**  
   - The author that the user most prefers.  
   - Calculated based on the author for whom the user gave the highest average rating.

2. **fav_publisher**  
   - The publisher that the user most prefers.  
   - Calculated as the publisher from which the user has rated the most books (highest count).

3. **avg_book_rate**  
   - The average rating for each book across all users.  
   - Provides a measure of overall book popularity/quality.

4. **weighted_score**  
   - An IMDb-style weighted rating for each book.  
   - Balances a book’s average rating with the global mean to prevent books with very few ratings from dominating the ranking.


# Converting to Pivot table

## Building User Preference Vectors

### Why?
* To use Collaborative Filtering, we need to represent each user’s preferences in a numerical form. This allows us to measure similarity between users (user-based CF) or between books (item-based CF).

### How?

* Construct a matrix User × Book.

* Rows = users, Columns = books (ISBN or Title).

* Values = ratings (or 0 if the user has not rated the book).

### Importance:

* These vectors allow us to calculate user similarity (e.g., cosine similarity).

* Enable recommendation: if two users are similar, recommend to one the books rated highly by the other.

user_item_matrix =  df_plus.pivot_table(
    index= "user_id",
    columns="title",
    values="book_rating",
    fill_value=0
)

user_item_matrix

user_item_matrix.shape

🔹 Sparse Matrix

* In the User–Item Matrix, most values are 0 because users rate only a few books.

* A dense matrix would be huge and waste memory.

* A sparse matrix stores only the non–zero values.

* It reduces memory usage and speeds up calculations (e.g., cosine similarity, matrix factorization).

* This makes collaborative filtering possible on large datasets.


sparse_user_item = csr_matrix(user_item_matrix.values)

sparse_user_item
